Make token price use big float instead of float32 (#16)
* Make token price a big float

* Use float64 instead of big.Float

* Remove copying migration folder in Dockerfile

+++ core/oracle/token_price_manager.go
@@ -18,7 +18,7 @@ type Response struct {
 	Data map[string]struct {
 		Quote struct {
 			Usd struct {
-				Value float32 `json:"price"`
+				Value float64 `json:"price"`
 			} `json:"USD"`
 		} `json:"quote"`
 	} `json:"data"`
